requirements.txt
pyautogui>=0.9.54
pydirectinput>=1.0.4
keyboard>=0.13.5
opencv-python>=4.8.0
numpy>=1.26.0
pynput>=1.7.6
pytesseract>=0.3.10
Pillow>=10.0.0